refactor[v1]: split monolithic router into focused modules
router.py → thin assembler (12 lines)
helpers.py → shared utilities (CountryCode enum, _energy_type_for, _resolve_*, etc.)
cache.py → warming logic + state dicts (_warm_all_v1_caches, now imported from here)
routes/capability.py → tag="Capability" (sources, countries, region-types, generation-sources)
routes/regions.py → tag="Regions" (list regions, get region)
routes/forecasts.py → tag="Forecasts" (per-region, snapshot, timeseries, refresh)
routes/generation.py → tag="Generation" (per-region, snapshot, timeseries, refresh)

Swagger UI now shows 4 labelled sections. All 32 tests pass.
